Use pandas testing methods instead of our clumsy ones.

+++ tests/trnsysGUI/diagram/_testHelper.py
@@ -1,10 +1,11 @@
 import dataclasses as _dc
 import pathlib as _pl
 import shutil as _su
-import subprocess as _sp
 import typing as _tp
 
 import pandas as _pd
+
+from . import _git
 
 DATA_DIR = _pl.Path(__file__).parent / "data"
 
@@ -33,23 +34,32 @@ class Helper:
         self,
         project: Project,
     ):
-        self._project = project
+        self._projectFolderPathInExamplesDir = self._getProjectFolderPathInExamplesDir(project)
 
-        testCasesFolderPath = DATA_DIR / self._project.testCasesDirName
+        testCasesFolderPath = DATA_DIR / project.testCasesDirName
 
         # project name is used twice because the project file must always be contained in a folder
         # of the same name
-        self.actualProjectFolderPath = testCasesFolderPath / self._project.projectName / self._project.projectName
+        self.actualProjectFolderPath = testCasesFolderPath / project.projectName / project.projectName
 
-        self._expectedProjectFolderPath = testCasesFolderPath / self._project.projectName / "expected"
+        self._expectedProjectFolderPath = testCasesFolderPath / project.projectName / "expected"
 
-    def setup(self):
-        if self._project.exampleDirNameToCopyFrom:
-            self._copyExampleToTestInputFolder()
+    @staticmethod
+    def _getProjectFolderPathInExamplesDir(project: Project) -> _tp.Optional[_pl.Path]:
+        if not project.exampleDirNameToCopyFrom:
+            return None
 
+        pytrnsysGuiDir = _pl.Path(__file__).parents[3]
+
+        exampleProjectPath = pytrnsysGuiDir / "data" / project.exampleDirNameToCopyFrom / project.projectName
+
+        return exampleProjectPath
+
+    def setup(self) -> None:
+        self._copyExampleToTestInputFolderIfNeeded()
         self._removeUntrackedPaths()
 
-    def _removeUntrackedPaths(self):
+    def _removeUntrackedPaths(self) -> None:
         untrackedPaths = self._getUntrackedPathsInActualProjectFolder()
 
         for untrackedPath in untrackedPaths:
@@ -58,80 +68,17 @@ class Helper:
             else:
                 untrackedPath.unlink()
 
-    def _getUntrackedPathsInActualProjectFolder(self):
+    def _getUntrackedPathsInActualProjectFolder(self) -> _tp.Sequence[_pl.Path]:
         projectFolderPathUnderVersionControl = self._getProjectFolderPathUnderVersionControl()
 
-        untrackedRelativePaths = self._getUntrackedRelativePaths(projectFolderPathUnderVersionControl)
+        untrackedRelativePaths = _git.getUntrackedRelativePaths(projectFolderPathUnderVersionControl)
 
         untrackedAbsolutePaths = [self.actualProjectFolderPath / p for p in untrackedRelativePaths]
 
         return untrackedAbsolutePaths
 
-    def _getProjectFolderPathUnderVersionControl(self):
-        if self._project.exampleDirNameToCopyFrom:
-            return self._getProjectFolderPathInExamplesDir()
-
-        return self.actualProjectFolderPath
-
-    def _getProjectFolderPathInExamplesDir(self):
-        assert self._project.exampleDirNameToCopyFrom
-
-        pytrnsysGuiDir = _pl.Path(__file__).parents[3]
-
-        exampleProjectPath = (
-            pytrnsysGuiDir / "data" / self._project.exampleDirNameToCopyFrom / self._project.projectName
-        )
-
-        return exampleProjectPath
-
-    def _getUntrackedRelativePaths(self, projectFolderPathUnderVersionControl):
-        untrackedPathsRelativeToRepoRoot = self._getUntrackedPathsRelativeToRepoRoot(
-            projectFolderPathUnderVersionControl
-        )
-
-        repositoryRootDirPath = self._getRepositoryRootDirPath()
-
-        untrackedAbsolutePaths = [repositoryRootDirPath / p for p in untrackedPathsRelativeToRepoRoot]
-
-        untrackedRelativePaths = [p.relative_to(projectFolderPathUnderVersionControl) for p in untrackedAbsolutePaths]
-
-        return untrackedRelativePaths
-
-    @staticmethod
-    def _getUntrackedPathsRelativeToRepoRoot(projectFolderPathUnderVersionControl):
-        completedProcess = _sp.run(
-            ["git", "status", "-unormal", "--porcelain", str(projectFolderPathUnderVersionControl)],
-            check=True,
-            text=True,
-            capture_output=True,
-        )
-
-        statusesAndUntrackedPath = completedProcess.stdout.splitlines()
-
-        untrackedPaths = []
-        for statusAndUntrackedPath in statusesAndUntrackedPath:
-            status, untrackedPath = statusAndUntrackedPath.split()
-
-            modifiedStatus = "??"
-            if status != modifiedStatus:
-                continue
-
-            untrackedPaths.append(untrackedPath)
-
-        return untrackedPaths
-
-    @staticmethod
-    def _getRepositoryRootDirPath():
-        completedProcess = _sp.run(
-            ["git", "rev-parse", "--show-toplevel"],
-            check=True,
-            text=True,
-            capture_output=True,
-        )
-
-        repositoryRootDirPath = completedProcess.stdout.splitlines()[0]
-
-        return _pl.Path(repositoryRootDirPath)
+    def _getProjectFolderPathUnderVersionControl(self) -> _pl.Path:
+        return self._projectFolderPathInExamplesDir or self.actualProjectFolderPath
 
     def ensureFilesAreEqual(
         self, fileRelativePathAsString: str, replaceInExpected: _tp.Optional[_tp.Tuple[str, str]] = None
@@ -147,33 +94,25 @@ class Helper:
 
         assert actualContent == expectedContent
 
-    def _getActualAndExpectedFilePath(self, fileRelativePathAsString):
+    def _getActualAndExpectedFilePath(self, fileRelativePathAsString: str) -> _tp.Tuple[_pl.Path, _pl.Path]:
         fileRelativePath = _pl.Path(fileRelativePathAsString)
         actualFilePath = self.actualProjectFolderPath / fileRelativePath
         expectedFilePath = self._expectedProjectFolderPath / fileRelativePath
         return actualFilePath, expectedFilePath
 
-    def ensureCSVsAreEqual(self, fileRelativePathAsString: str, absoluteTolerance: float = 1e-10) -> None:
+    def ensureDataFramesAreEqual(self, fileRelativePathAsString: str) -> None:
         actualFilePath, expectedFilePath = self._getActualAndExpectedFilePath(fileRelativePathAsString)
 
         actualDf: _pd.DataFrame = _pd.read_csv(actualFilePath, delim_whitespace=True)
         expectedDf: _pd.DataFrame = _pd.read_csv(expectedFilePath, delim_whitespace=True)
 
-        assert actualDf.shape == expectedDf.shape
+        _pd.testing.assert_frame_equal(actualDf, expectedDf, atol=1e-10)
 
-        actualColumns = "\n".join(sorted(actualDf.columns))  # pylint: disable=no-member
-        expectedColumns = "\n".join(sorted(expectedDf.columns))  # pylint: disable=no-member
+    def _copyExampleToTestInputFolderIfNeeded(self) -> None:
+        if not self._projectFolderPathInExamplesDir:
+            return
 
-        assert actualColumns == expectedColumns
-
-        maxAbsoluteDifference = (actualDf - expectedDf).max().max()
-
-        assert maxAbsoluteDifference <= absoluteTolerance
-
-    def _copyExampleToTestInputFolder(self):
         if self.actualProjectFolderPath.exists():
             _su.rmtree(self.actualProjectFolderPath)
 
-        projectPathInExamplesDir = self._getProjectFolderPathInExamplesDir()
-
-        _su.copytree(projectPathInExamplesDir, self.actualProjectFolderPath)
+        _su.copytree(self._projectFolderPathInExamplesDir, self.actualProjectFolderPath)
